Share bytecode IR between CPU and GPU trace gen
Extract the stack-machine bytecode (OpCode / ExprSpan / BusMeta + the
AST-walker emit_expr / compile_bus_to_bytecode) into a new
feature-agnostic module openvm/src/bytecode.rs. cuda_abi re-exports the
types for back-compat; cuda/mod.rs keeps compile_derived_to_gpu but
delegates bus-interaction compilation to the shared module.

Add a CPU eval_expr interpreter that mirrors openvm/cuda/src/expr_eval.cuh:
fixed-size [F; 16] stack on the local frame, unsafe get_unchecked
indexing, single dispatch loop. Mirrors the GPU layout exactly:
Vec<u32> bytecode + ExprSpan{off,len} indexing into it.

Wire cpu/mod.rs to call compile_bus_to_bytecode(apc_height=1) so the
PushApc operand is the column index directly, then iterate the shared
bytecode/spans per bus interaction. Removes the per-eval Vec allocation
that the prior Vec<F> stack form had.

Drop the previous polynomial-form fast-eval (CompiledExpr / Polynomial /
decompose / mul_decompositions / CompiledBusInteraction) from
autoprecompiles/src/expression.rs - about 200 lines superseded by the
shared bytecode path.

Per keccak/10K/APC=30/mock: Polynomial 5.05/5.09/2.34s vs shared
bytecode 5.29/5.41/2.50s (+5-6%). In exchange CPU and GPU share one
compile pipeline, runtime is degree-agnostic (Polynomial was capped at
2), and ~200 LoC of duplicate AST decomposition logic is gone.
