+++ openvm/src/powdr_extension/trace_generator/cpu/mod.rs
@@ -153,24 +153,15 @@ impl<ISA: OpenVmISA> PowdrTraceGeneratorCpu<ISA> {
             &self.apc,
         );
 
-        // Build dense Vec indexed by poly ID for O(1) column lookups in the hot loop.
-        // Poly IDs may be sparse (gaps between IDs), so the Vec is sized to max_id + 1.
         let width = apc_poly_id_to_index.len();
-        let max_poly_id = apc_poly_id_to_index.keys().last().copied().unwrap_or(0) as usize;
-        let apc_poly_id_to_index: Vec<usize> = (0..=max_poly_id)
-            .map(|id| apc_poly_id_to_index.get(&(id as u64)).copied().unwrap_or(0))
-            .collect();
 
-        // Compile bus interactions once before the hot loop
-        let compiled_interactions = {
-            use powdr_autoprecompiles::expression::CompiledBusInteraction;
-            CompiledBusInteraction::compile_all(
-                &self.apc.machine().bus_interactions,
-                &apc_poly_id_to_index,
-                BabyBear::ZERO,
-                BabyBear::ONE,
-            )
-        };
+        // Compile bus interactions to shared bytecode (apc_height = 1 so the
+        // PushApc operand is the column index directly).
+        let (bus_meta, arg_spans, bytecode) = crate::bytecode::compile_bus_to_bytecode(
+            &self.apc.machine().bus_interactions,
+            &apc_poly_id_to_index,
+            1,
+        );
 
         // allocate for apc trace
         let height = next_power_of_two_or_zero(num_apc_calls);
@@ -198,35 +189,45 @@ impl<ISA: OpenVmISA> PowdrTraceGeneratorCpu<ISA> {
                 // Fill in the columns we have to compute from other columns
                 // (these are either new columns or for example the "is_valid" column).
                 for derived_column in columns_to_compute {
-                    let col_index = apc_poly_id_to_index[derived_column.variable.id as usize];
+                    let col_index = apc_poly_id_to_index[&derived_column.variable.id];
                     row_slice[col_index] = match &derived_column.computation_method {
                         ComputationMethod::Constant(c) => *c,
                         ComputationMethod::QuotientOrZero(e1, e2) => {
                             use powdr_number::ExpressionConvertible;
 
                             let divisor_val = e2.to_expression(&|n| *n, &|column_ref| {
-                                row_slice[apc_poly_id_to_index[column_ref.id as usize]]
+                                row_slice[apc_poly_id_to_index[&column_ref.id]]
                             });
                             if divisor_val.is_zero() {
                                 BabyBear::ZERO
                             } else {
                                 divisor_val.inverse()
                                     * e1.to_expression(&|n| *n, &|column_ref| {
-                                        row_slice[apc_poly_id_to_index[column_ref.id as usize]]
+                                        row_slice[apc_poly_id_to_index[&column_ref.id]]
                                     })
                             }
                         }
                     };
                 }
 
-                // Evaluate bus interactions using compiled expressions.
-                // Periphery chips use AtomicU32 counters — thread-safe.
-                for ci in &compiled_interactions {
-                    let mult = ci.mult.eval(row_slice);
+                // Evaluate bus interactions via shared stack-machine interpreter
+                // (same bytecode IR as the GPU kernel in `expr_eval.cuh`).
+                for meta in &bus_meta {
+                    let base = meta.args_index_off as usize;
+                    let mult_span = arg_spans[base];
+                    let mult = crate::bytecode::eval_expr(&bytecode, mult_span, row_slice);
+                    let num_args = meta.num_args as usize;
                     periphery_real.apply(
-                        ci.id as u16,
+                        meta.bus_id as u16,
                         mult.as_canonical_u32(),
-                        ci.args.iter().map(|a| a.eval(row_slice).as_canonical_u32()),
+                        (0..num_args).map(|i| {
+                            crate::bytecode::eval_expr(
+                                &bytecode,
+                                arg_spans[base + 1 + i],
+                                row_slice,
+                            )
+                            .as_canonical_u32()
+                        }),
                         periphery_bus_ids,
                     );
                 }
